Finished read from input stream

+++ src/main/java/com/hobbes/wstore/DiskDataRange.java
@@ -18,10 +18,8 @@ class DiskDataRange extends DataRange {
     }
 
     @Override
-    public long getData(long relativeStartPosition, byte[] buf, int pos, int len) throws IOException {
-	// Do nothing for now
-		return ((long)(in.read(relativeStartPosition, buf, pos, len)));
-
+    public int getData(long relativeStartPosition, byte[] buf, int pos, int len) throws IOException {
+		return in.read(relativeStartPosition, buf, pos, len);
     }
 
     public DataRange getSubrange(long start, long end) {
